requirements.txt
faiss-cpu==1.7.4
fitz==0.0.1.dev2          # pymupdf
numpy==1.26.4
sentence-transformers==2.2.2
streamlit==1.35.0
requests==2.32.3
